fixed g++ errors, removed unused code

+++ CachedPNGWriter.cpp
@@ -197,7 +197,7 @@ bool CachedPNGWriter::compose(const std::string& path, const double scale)
 		for (auto it = m_partList.begin(); it != m_partList.end(); ++it) {
 			ImagePart& img = *it;
 			// do we have to open this image?
-			if (img.y != y && img.pngPtr == nullptr) {
+			if (img.y != static_cast<int>(y) && img.pngPtr == nullptr) {
 				continue;   // Not your turn, image!
 			}
